create debug test runner that creates a test container for debugging tests
