+++ packages/vscode-extension/e2e_tests/testRunnerDebug.ts
@@ -1,0 +1,7 @@
+import { createAndStartTestContainer } from './setupTestContainer';
+import { run as testRunner } from './testRunner';
+
+export async function run(): Promise<void> {
+  await createAndStartTestContainer();
+  return testRunner();
+}
